pe-nodejs source is arch-dependent
We're building pe-nodejs by downloading pre-built nodejs binaries and
packaging them for debian.  So, in the pe-nodejs the "source" is
a tarball of pre-built binaries for a particular architecture.  So, when
downloading the source you should specify the architecture.  If you
don't specify the arch, then when the build step happens, the downloaded
pre-built binary tarball arch and the build target arch might not match.