download sources to package-specific subdir
Commit d110a40da8a929649a854738497aa13e0378c6e5 allows a developer
to iterate by modifying sources in the download folder, but this
feature created a bug if two packages download the same component
at a different SHA or tag.  In that case, what happened is that
the first package downloaded the source at the SHA specified in its
build.sh, but then the second package failed to checkout the version
specified in its own build.sh since the component had already been
cloned by the first package.

This issue caused the build to break when building
mbed-edge-core-devmode after mbed-edge-core because the patch files
for mbed-edge-core-devmode didn't apply to the different version of the
edge-core component cloned by mbed-edge-core package.  The interim
fix can be seen in 7cf4fdaf4514889dbcd4c854c2ed68c1f83e9b50, in which
the version of edge-core cloned by mbed-edge-core is made to be the
same as mbed-edge-core-devmode, but it seems reasonable to allow
multiple packages to clone the same component at different revisions.

To fix this issue while still allowing a developer to modify local
source for dev/debug, sources are downloaded to a package-specific
subdirectory at
PELION_SOURCE_DIR/PELION_PACKAGE_NAME/PELION_PACKAGE_COMPONENT.