Fix Mac build authentication
Docker for Mac has a particular, non-standard way to forward the ssh
agent socket, by using a hard coded magic path.

This commit implements that, and also adds a fix so non-root users can
use the socket as well (since permissions and ownership of the socket
are incorrectly mapped by Docker for Mac).

Signed-off-by: Cristian Prundeanu <cristian.prundeanu@arm.com>