upgrade mbed-edge-core packages to 0.15.0
Adds GRM feature which allows a client to add new LwM2M resources
to a gateway device

Added the CMake flags related to new FOTA library to indicate that
new FOTA framework library is not compilied in Edge Core.

All patches were either updated for 0.15.0 to adjust to new APIs
or were removed because the patch content was upstreamed.

fix mbed-edge-core-devmode version