mbed-edge-core: add platform version file
This adds two LwM2M Resources intended for reporting platform versions.
The contents of each resource is read from a corresponding file on disk,
and it's up to the user to decide what values to write to the file.

It is intended for one file to contain a "friendly" or "readable" version
string, perhaps something that follows semantic versioning, and for the
other file to contain a hash over some data, such as the md5 hash of
a list of all installed packages.

Platform Version friendly version:
	file: /var/lib/pelion/readable_version
	lwm2m: /10252/0/10

Platform Version MD5 hash:
	file: /var/lib/pelion/platform_version
	lwm2m: /10252/0/11

If the file doesn't exist, mbed-cloud-client supplies a default
value of "-1" for the LwM2M resource.