Change containernetworking to c2d variant
Kubelet was reporting the following errors without using the c2d variant:
2145 cni.go:188] Unable to update cni config: No networks found in /etc/cni/net.d
2145 kubelet.go:2181] Container runtime network not ready: NetworkReady=false reason:NetworkPluginNotReady message:docker: network plugin is not ready: cni config uninitialized

This was due to a name clash where we were creating a containernetworking plugin
but the containernetworking plugin already existed.

The solution was to rename our package to containernetworking-c2d due to the
fact that our c2d plugin had c2d networking files. Then, install both our
containernetworking-c2d as well as the existing containernetworking

Note: Changed kubelet and k3s to use the new plugin