Modified maestro config to add GRM feature
This allows Maestro to add new LwM2M resources to gateway device. These
resources represents the gateway capabilities, meaning the rich node
services which are running on the gateway, for example - KaaS, Logging
and Terminal. Maestro also pushes the configuration data which is
associated to the services.

Remote configuration management via LwM2M -
With this feature, user can update the configuration of the service like
Maestro logs, KaaS or Terminal via LwM2M. Maestro on receiving this write
request will update the configuration file on the gateway.

TODO: How to restart the service after config file is updated?