fix buster ARM64 containernetworking-plugins install path
The dh-golang debhelper package is used to automate the build
of containernetworking-plugins.  The current Buster version
dh-golang 1.39 places build output in an arch-specific directory
which causes the plugin binaries to be installed into the wrong place.

Buster-Backports installs dh-golang version 1.15 which instead
builds in a common directory and fixes the installation path.