Backport from master (#224)
* remove support for CentOS/8 since it is EOL

CentOS/8 has reached End-Of-Life:
    https://www.centos.org/centos-linux-eol/

We're doing this now rather than later because CentOS can not be
upgraded with 'dnf update' or 'yum update' which obviously breaks our
build.

Here is a link to a centos forum post discussing this error and noting that
it WILL NOT be fixed by Redhat and users must instead migrate to a different
distribution:

    https://forums.centos.org/viewtopic.php?f=54&t=78708

Here is the error:
-----------------------------------------------------------------------
Error: Failed to download metadata for repo 'appstream': Cannot prepare internal mirrorlist: No URLs in mirrorlist
The command '/bin/sh -c yum update -y &&        yum install -y createrepo_c yum-utils sudo dnf-plugins-core epel-release &&     yum groupinstall -y 'Development Tools' &&     yum config-manager --set-enabled powertools &&     yum makecache' returned a non-zero code: 1

Error: Process completed with exit code 1.
-----------------------------------------------------------------------

* fix bionic build

install dwz package from bionic-backports so that it matches
the debhelper package also installed from bionic-backports.

* upgrade debian 10 docker image to buster-20220228

Let's upgrade the base image to pick up some CVE fixes.

* upgrade all to golang-1.16 except for kubelet

kubelet uses the build script provided by the kubelet project
which currently uses golang-1.11.  If we want to compile
kubelet with a different version of golang, we'll need to
customize our own scripts.

* kubelet: don't use dockerized make

the dockerized make environment comes with its own built-in
version of Golang, but we need to control the golang
version so that it matches the rest of Pelion Edge.

* delete devicedb package

DeviceDB is not now and was never actually used in Pelion Edge.

* fix edge-proxy github org

edge-proxy was moved from ArmPelionEdge to PelionIoT

* upgrade containernetworking/plugins to v0.8.5

Both the Yocto Meta-Pelion-Edge and the Snap-Pelion-Edge builds
are using this version.

* upgrade maestro-shell to v2.8.0

maestro-shell removed suport for grease which removes several
dependencies.

* containernetworking-plugins: fix CVE-2021-33194

bump golang.org/x/net version to fix CVE-2021-33194

* patch golang to fix CVE-2021-29923

Build a patched version Golang-1.16.15 to compile Pelion Edge
components.  This fixes CVE-2021-29923.

The Golang patch that fixes the CVE is backported from:
	https://go-review.googlesource.com/c/go/+/325829/

* upgrade kubelet to fix CVE-2021-33194

* upgrade edge-proxy to fix CVE-2021-33194

* upgrade maestro to fix CVE-2021-33194

* fix buster ARM64 containernetworking-plugins install path

The dh-golang debhelper package is used to automate the build
of containernetworking-plugins.  The current Buster version
dh-golang 1.39 places build output in an arch-specific directory
which causes the plugin binaries to be installed into the wrong place.

Buster-Backports installs dh-golang version 1.15 which instead
builds in a common directory and fixes the installation path.

* Corrected merge from master.

Co-authored-by: Nic Costa <nic.costa@gmail.com>